+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Fix documentation typo. Thanks @mark-summerfield.
 - Add more to radio button docs.
 - Fix BmpImage::from_data() not passing the correct len.
+- HelpDialog refactoring, add HelpDialog::shown().
+- Pull FLTK IME fixes.
 
 ## [1.2.20] - 2021-12-09
 - Cleanup macros and surrounding api. Thanks @AshfordN.